Refactor Makefile and Tests
This commit refactors the Makefile and Tests files. It renames the target "testsNoLog" to "testNoLog" in the Makefile. Additionally, it removes the duplicate "make cleanTests" command in the "testsNoLog" target. In the Tests directory, it removes the error messages from the "result-False.txt" file. Finally, it removes the unnecessary assertion in the "TestPDFValidity" class in the "tests.py" file.